refactor: share negative styling helper
