+++ ui/scan.py
@@ -4,15 +4,7 @@ from pandas.io.formats.style import Styler
 from utils.formatting import _bold, _usd, _pct, _safe
 from utils.scan import safe_run_scan
 from .history import _apply_dark_theme
-
-
-def _style_negatives(df: pd.DataFrame) -> Styler:
-    """Return a Styler adding class "neg" to negative numeric cells."""
-    classes = pd.DataFrame("", index=df.index, columns=df.columns)
-    num_cols = df.select_dtypes(include="number").columns
-    for col in num_cols:
-        classes.loc[df[col] < 0, col] = "neg"
-    return df.style.set_td_classes(classes)
+from .table_utils import _style_negatives
 
 
 def build_why_buy_html(row: dict) -> str:
